Relocate cec frame handling code
Move and encapsulate the cec frame handling code into the new module(s), significantly clarifying the functionality of the cec-task module.
